Correct many failings
The contact and portfolio pages had the jumbotron outside the container,
which made them look different from the home page. This hasnow been
fixed.
Sticky footers are turned off if the viewport isn't large enough to
contain the whole content plus the fixed footer. This is done with
media queries for each page.